move cropduster from lib to src dir
